DatasetTranslation/requirements.txt
streamlit
pandas
deepl
langdetect
python-dotenv
matplotlib
plotly